refactor into single typechecker pass, use checked recursion
